+++ include/pbbam/PbiFile.h
@@ -40,8 +40,9 @@ struct PbiFile
         Version_3_0_0 = 0x030000,  ///< v3.0.0
         Version_3_0_1 = 0x030001,  ///< v3.0.1
         Version_3_0_2 = 0x030002,  ///< v3.0.2
+        Version_4_0_0 = 0x040000,  ///< v4.0.0
 
-        CurrentVersion = Version_3_0_2  ///< Synonym for the current PBI version.
+        CurrentVersion = Version_4_0_0  ///< Synonym for the current PBI version.
     };
 
     ///
@@ -73,7 +74,9 @@ struct PbiFile
         N_INS,
         N_DEL,
         MAP_QUALITY,
-        STRAND
+        STRAND,
+        N_INS_OPS,
+        N_DEL_OPS,
     };
 
     ///