+++ tests/src/test_BamRecord.cpp
@@ -2805,4 +2805,84 @@ TEST(BAM_BamRecord, fetches_holenumber_from_name_ccs)
     EXPECT_THROW(bam.HoleNumber(), std::runtime_error);
 }
 
+TEST(BAM_BamRecord, can_determine_num_inserted_bases)
+{
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "4=3I4D1=2I4=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(5, record.NumInsertedBases());
+    }
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "14=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(0, record.NumInsertedBases());
+    }
+}
+
+TEST(BAM_BamRecord, can_determine_num_deleted_bases)
+{
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "4=3I4D1=2I4=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(4, record.NumDeletedBases());
+    }
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "14=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(0, record.NumDeletedBases());
+    }
+}
+
+TEST(BAM_BamRecord, can_determine_num_insertion_operations)
+{
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "4=3I4D1=2I4=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(2, record.NumInsertionOperations());
+    }
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "14=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(0, record.NumInsertionOperations());
+    }
+}
+
+TEST(BAM_BamRecord, can_determine_num_deletion_operations)
+{
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "4=3I4D1=2I4=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(1, record.NumDeletionOperations());
+    }
+    {
+        const auto record = BamRecordTests::MakeCigaredRecord(
+            "GATTACAGATTACA",
+            "14=",
+            Data::Strand::FORWARD
+        );
+        EXPECT_EQ(0, record.NumDeletionOperations());
+    }
+}
+
 // clang-format on